whoops wrong type

+++ src/main/java/canaryprism/slavacord/data/SlashCommandOptionChoiceData.java
@@ -20,7 +20,7 @@ public record SlashCommandOptionChoiceData<T>(
 
     public canaryprism.discordbridge.api.data.interaction.slash.SlashCommandOptionChoiceData toSlashCommandOptionChoiceBuilder() {
         var value = (this.value instanceof Enum<?> enum_value) ?
-                ((Integer) enum_value.ordinal())
+                ((long) enum_value.ordinal())
                 : this.value;
         var builder = new canaryprism.discordbridge.api.data.interaction.slash.SlashCommandOptionChoiceData(name, value);
